Refactoring files names and clean up
