+++ src/pyvisa_proxy/_main.py
@@ -10,7 +10,7 @@ import typing
 from atexit import register
 from weakref import WeakMethod
 
-from .ProxyServer import ProxyServer
+from .proxy_server import ProxyServer
 
 LOGGER = logging.getLogger(__name__)
