+++ src/pyvisa_proxy/_version_handling.py
@@ -4,14 +4,13 @@
 :license: MIT, see LICENSE for more details.
 """
 
-import sys
 from importlib.metadata import PackageNotFoundError, version
 
 
 def get_version():
     """Retrieve the package version."""
     try:
-        return version("pyvisa-proxy")
+        return version(__package__)
     except PackageNotFoundError:
         # package is not installed
         return "unknown"
