+++ src/pyvisa_proxy/proxy_resource.py
@@ -8,7 +8,7 @@ import typing
 
 from pyvisa import Resource
 
-from .RpcClient import RpcClient
+from .rpc_client import RpcClient
 
 
 class ProxyResource(Resource):
